requirements.txt
streamlit           
groq               
tavily-python      
python-dotenv       
pytesseract       
easyocr         
opencv-python     
pillow            
torch              
transformers    